changement des double en float

+++ src/Odometry.cpp
@@ -49,7 +49,7 @@ void Odometry::refresh(float m_encoderDeltaRight_mm, float m_encoderDeltaLeft_mm
      */
     m_deltaDist = (m_encoderDeltaLeft_mm + m_encoderDeltaRight_mm) / 2;
     float diffCount = m_encoderDeltaRight_mm - m_encoderDeltaLeft_mm;
-    m_deltaTheta = diffCount / double(m_encoderWheelsDistance_mm); // En radian
+    m_deltaTheta = diffCount / m_encoderWheelsDistance_mm; // En radian
 
     if (diffCount == 0)    // On considère le mouvement comme une ligne droite
     {
