+++ src/Odometry.h
@@ -54,8 +54,8 @@ private:
 
     float m_encoderWheelsDistance_mm;
 
-    double m_deltaTheta;
-    double m_deltaDist;
+    float m_deltaTheta;
+    float m_deltaDist;
 
     // Position actuelle
     float m_X_mm, m_Y_mm; // En mm
